requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
astroid=2.4.0=py37_0
backcall=0.1.0=py37_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2020.1.1=0
cairo=1.14.12=h8948797_3
certifi=2020.4.5.1=py37_0
cffi=1.14.0=py37h2e261b9_0
click=7.1.2=py_0
cuda100=1.0=0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=py37_0
dbus=1.13.14=hb2f20db_0
decorator=4.4.2=py_0
entrypoints=0.3=py37_0
expat=2.2.6=he6710b0_0
ffmpeg=4.0=hcdf2ecd_0
flask=1.1.2=py_0
flask-dropzone=1.5.4=pypi_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
gevent=1.4.0=py37h7b6447c_0
glib=2.63.1=h5a9c865_0
graphite2=1.3.13=h23475e2_0
greenlet=0.4.15=py37h7b6447c_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
gunicorn=19.8.0=py37_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
intel-openmp=2020.0=166
ipykernel=5.1.4=py37h39e3cac_0
ipython=7.13.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
isort=4.3.21=py37_0
itsdangerous=1.1.0=py37_0
jasper=2.0.14=h07fcdf6_1
jedi=0.16.0=py37_1
jinja2=2.11.2=py_0
jpeg=9b=h024ee3a_2
jupyter_client=6.1.2=py_0
jupyter_core=4.6.3=py37_0
kiwisolver=1.1.0=py37he6710b0_0
lazy-object-proxy=1.4.3=py37h7b6447c_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
markupsafe=1.1.1=py37h7b6447c_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37hef1b27d_0
mccabe=0.6.1=py37_1
mkl=2020.0=166
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
ncurses=6.2=he6710b0_0
ninja=1.9.0=py37hfd86e86_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openssl=1.1.1g=h7b6447c_0
parso=0.6.2=py_0
pcre=8.43=he6710b0_0
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pillow=7.0.0=py37hb39fc2d_0
pip=20.0.2=py37_1
pixman=0.38.0=h7b6447c_0
prompt-toolkit=3.0.4=py_0
prompt_toolkit=3.0.4=0
ptyprocess=0.6.0=py37_0
py-opencv=3.4.2=py37hb342d67_1
pycparser=2.20=py_0
pygments=2.6.1=py_0
pylint=2.5.2=py37_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.7=hcf32534_0_cpython
python-dateutil=2.8.1=py_0
pytorch=1.5.0=py3.7_cuda10.1.243_cudnn7.6.3_0
pytorch-cpu=1.1.0=py3.7_cpu_0
pyzmq=18.1.1=py37he6710b0_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
setuptools=46.1.3=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.14.0=py37_0
sqlite=3.31.1=h7b6447c_0
tk=8.6.8=hbc83047_0
toml=0.10.0=py37h28b3542_0
torchvision=0.6.0=py37_cu101
tornado=6.0.4=py37h7b6447c_1
traitlets=4.3.3=py37_0
typed-ast=1.4.1=py37h7b6447c_0
wcwidth=0.1.9=py_0
werkzeug=1.0.1=py_0
wheel=0.34.2=py37_0
wrapt=1.11.2=py37h7b6447c_0
xz=5.2.4=h14c3975_4
zeromq=4.3.1=he6710b0_3
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
